Refactoring: Extracted the job handling from the inner class into a dedicated class, and renamed it properly
